requirements.txt
aiogram==3.13.0
apscheduler==3.10.4
pytz
psycopg2-binary